Update createk8scluster.md
The convention is to use <> to surround required input in shell commands.  
I had to explain the command several times when holding the workshop, as my attendees ran into issues around the automatic creation of the service principal for the AKS cluster.  
I believe that this small change will go a long way to making this easier to understand for some.